feat(io): move SpectraShop parser to feature-gated module, add copyright field
- Extract all ss_* helpers and SsMeta into io/src/spectrashop.rs, gated
  behind the new `spectrashop` Cargo feature (on by default)
- Add integration tests in io/tests/spectrashop_data.rs (gated with
  #![cfg(feature = "spectrashop")])
- Add `copyright: Option<String>` to SpectrumMetadata; update JSON schema
  and module-doc table accordingly
- Add licensing/format-spec prose to the SpectraShop module doc and README
- Fix non-UTF-8 SpectraShop files (lossy read) and embedded-quote sample
  names (tab-split tokenizer)
